+++ NEWS.md
@@ -11,6 +11,11 @@
   * Add a "Version" custom property and handle it in all RomData subclasses
     that have a "Version" or "Revision" field.
 
+* Bug fixes:
+  * Dreamcast and SegaSaturn incorrectly added DiscNumber as integer.
+    They should be uint. The end result is these properties didn't
+    show up correctly.
+
 ## v2.8 (released 2026/06/20)
 
 * New parsers: